more casts
git-svn-id: https://svn.r-project.org/R/trunk@59121 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/builtin.c
@@ -622,7 +622,7 @@ SEXP attribute_hidden do_cat(SEXP call, SEXP op, SEXP args, SEXP rho)
 			strncpy(buf, p, 512); buf[511] = '\0';
 			p = buf;
 		    }
-		    w = strlen(p);
+		    w = (int) strlen(p);
 		    cat_sepwidth(sepr, &sepw, ntot);
 		    /* This is inconsistent with the version above.
 		       As from R 2.3.0, fill <= 0 is ignored. */
